+++ src/main/java/edu/chalmers/notenoughspace/core/entity/beamable/Junk.java
@@ -23,8 +23,8 @@ public class Junk extends BeamableEntity {
         return 1f;
     }
 
-    public float getPoints() {
-        return 0f;
+    public int getPoints() {
+        return 0;
     }
 
     public void update(float tpf) {
